cleaned up code. transferred some code to a library.
